Converts float to double (#2343)
* Limiting the scope of variables #874

Limited the scope of variables in moveit_core/collision_detection

* Update moveit_core/collision_detection/src/collision_octomap_filter.cpp


* Update moveit_core/collision_detection/src/collision_octomap_filter.cpp


* Update moveit_core/collision_detection/src/collision_octomap_filter.cpp


* convert float to double

* change double to float

* Feedback fixes

* Introduced variables removed from previous merge commit

* Updated GL_Renderer function definitions with double instead of float

* Changed update() function arguments to float since it is a derived virtual function and needs to be overriden

* Fixed all override errors in visualization

* *Fixed override errors in perception
*Changed reinterpret_cast to double* from float*

* change variable types to fit function definition

* Fixed clang-tidy warnings

* Fixed scope of reusable variables

---------

+++ moveit_core/planning_scene/src/planning_scene.cpp
@@ -1058,7 +1058,7 @@ bool PlanningScene::loadGeometryFromStream(std::istream& in, const Eigen::Isomet
           RCLCPP_ERROR(LOGGER, "Failed to read pose from scene file");
           return false;
         }
-        float r, g, b, a;
+        double r, g, b, a;
         if (!(in >> r >> g >> b >> a))
         {
           RCLCPP_ERROR(LOGGER, "Improperly formatted color in scene geometry file");
